[BEAM-1689] Apply changes for Flink's StatefulDoFnRunner to the primary StatefulDoFnRunner
